IPRT/VISO: Pass a the directory of a VISO file to the ISO maker to allow for relative file/dir specification from it. Nested push-iso now possible.  Tweak importer code to handle NT 3.1 iso where the big endian volume sequence number wasn't set. [build ifx]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@70398 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/fs/isomakercmd.cpp
@@ -2084,7 +2084,7 @@ static int rtFsIsoMakerCmdAddSomething(PRTFSISOMAKERCMDOPTS pOpts, const char *p
 static int rtFsIsoMakerCmdOptPushIso(PRTFSISOMAKERCMDOPTS pOpts, const char *pszIsoSpec, const char *pszOption, uint32_t fFlags)
 {
     int32_t iSrcStack = pOpts->iSrcStack + 1;
-    if (iSrcStack >= RT_ELEMENTS(pOpts->aSrcStack))
+    if ((uint32_t)iSrcStack >= RT_ELEMENTS(pOpts->aSrcStack))
         return rtFsIsoMakerCmdErrorRc(pOpts, VERR_NOT_IMPLEMENTED,
                                       "Too many pushes %s %s (previous: %s %s, %s %s, %s %s, ...)",
                                       pszOption, pszIsoSpec,
